requirements.txt
chainlit>=2.0.0
crewai==0.5.0
langchain==0.1.0
langchain-openai==0.0.2
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
linkedin-api==2.0.0
openai>=1.7.1,<2.0.0
tiktoken>=0.5.2,<0.6.0
aiofiles==23.2.1
pydantic==2.5.0 